ready to Controller
