+++ src/main/java/mission/repository/ChannelRepository.java
@@ -3,14 +3,14 @@ package mission.repository;
 import mission.entity.Channel;
 
 import java.io.IOException;
-import java.util.List;
+import java.util.Optional;
 import java.util.Set;
 import java.util.UUID;
 
 public interface ChannelRepository {
-    Channel register(Channel channel) throws IOException;
+    Channel create(Channel channel) throws IOException;
     Set<Channel> findAll() throws IOException;
-    Channel findById(UUID id) throws IOException, ClassNotFoundException;
+    Optional<Channel> findById(UUID id) throws IOException, ClassNotFoundException;
     //Channel updateChannelName(Channel updatingChannel);
     void deleteById(UUID id);
 }
